Missing IRun, set Judgement defaults to null

+++ ContestModel/src/org/icpc/tools/contest/model/IRun.java
@@ -37,7 +37,7 @@ public interface IRun extends IContestObject {
 	 *
 	 * @return the contest time
 	 */
-	int getContestTime();
+	long getContestTime();
 
 	/**
 	 * Returns the wall clock time, in ms since the epoch.
